test: E2E tests for rescue files
